Make sure we have a hash
FossilOrigin-Name: 54fecc7fd3181ea107ee43a9d3e5e2a2d7ffd590bb009da2a128083195e0d9a4

+++ src/include/lib/Paheko/Entities/Files/FileThumbnailTrait.php
@@ -23,11 +23,16 @@ trait FileThumbnailTrait
 			return;
 		}
 
+		if (!isset($this->md5)) {
+			return;
+		}
+
 		// clean up thumbnails
 		foreach (self::ALLOWED_THUMB_SIZES as $size => $operations) {
 			Static_Cache::remove(sprintf(self::THUMB_CACHE_ID, $this->md5, $size));
 
 			$uri = $this->thumb_uri($size, false);
+
 			if ($uri) {
 				Web_Cache::delete($uri);
 			}
